% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-actionButtonModule.R
\name{actionButtonUI}
\alias{actionButtonUI}
\title{UI function of actionButton}
\usage{
actionButtonUI(id, label, ...)
}
\arguments{
\item{id}{module id}

\item{label}{label}

\item{...}{further arguments passed to input}
}
\description{
UI function of actionButton
}
