% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plotModule.R
\name{addTitleToPNG}
\alias{addTitleToPNG}
\title{Add title to PNG image}
\usage{
addTitleToPNG(path, title_format)
}
\arguments{
\item{path}{path to the image}

\item{title_format}{list with title text and format arguments}
}
\description{
Add title to PNG image
}
