% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_coords.R
\name{boundaries_coords.df}
\alias{boundaries_coords.df}
\title{Retrieve a list of boundaries' spacial mercator coordinates for a list of boundaries lat/long coordinates.}
\usage{
boundaries_coords.df(data, map_center = NULL)
}
\arguments{
\item{data}{List of Longitude/latitude coordinates of the boundaries. The Dataset is a data.frame containing a set of three columns (id, long, lat). 'id' is a unique representation of the boundary(ex: ZCTA, FSA). 'id' type is character.}

\item{map_center}{Spacial coordinates of a map center. The haversine distance in miles is computed if map_center is defined.}
}
\value{
list of spacial coordinates(lat/long and mercator) and a distance as data.frame
}
\description{
Retrieve a list of mercator boundaries with their spacial coordinates and a computed haversine distance in miles from the map center. Boundaries are defined in a shapefile dataframe(see method: shape_boundaries) as a list of vertices coordinates. The boundaries spacial coordinates is the center of gravity of its vertices.
}
\examples{
data = data.frame(long = c(-94.823686, -94), lat = c(39.121296, 39) )
map_center = data.frame(long = c(-94.823686), lat = c(39.121296) )
boundaries_coords<-boundaries_coords.df(data, map_center)

}
