% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_coords.R
\name{boundaries_coords.shape}
\alias{boundaries_coords.shape}
\title{Retrieve a list of boundaries' spacial coordinates(lat/long and mercator) from a shapefile}
\usage{
boundaries_coords.shape(shape_boundaries, map_center = NULL)
}
\arguments{
\item{shape_boundaries}{List of vertices coordinates as dataframe. Each boundaries is defined as a set of vertices. The value is given using shape_boundaries method}

\item{map_center}{Spacial coordinates of a map center. The haversine distance in miles is computed if map_center is defined.}
}
\value{
list of spacial coordinates(lat/long and mercator) and a distance as data.frame
}
\description{
Retrieve a list of boundaries with their spacial coordinates(lat/long and mercator) and a computed haversine distance in miles from the map center. Boundaries are defined in a shapefile dataframe(see method: shape_boundaries) as a list of vertices coordinates. The boundaries spacial coordinates is the center of gravity of its vertices.
}
\examples{
shape_boundaries<-shape_boundaries(shape_boundaries_dir, shape_boundaries_filename,boundary_identifier)
map_center = data.frame(long = c(-94.823686), lat = c(39.121296) )
boundaries_coords<-boundaries_coords.shape(shape_boundaries, map_center)

}

