% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_coords.R
\name{boundaries_coords_filter}
\alias{boundaries_coords_filter}
\title{Remove the overlapping boundaries}
\usage{
boundaries_coords_filter(boundaries_coords, distance)
}
\arguments{
\item{boundaries_coords}{List of boundaries and coordinates as a dataframe (id, lat, long)}

\item{distance}{distance in miles. If two boundaries are within 'distance', then one of the two boundaries is filtered out.}
}
\value{
List of boundaries and coordinates as a dataframe
}
\description{
Dirty way to detect overlapping boundaries and only return one. Used in boundary label creation in heatmap.
}
\examples{

}

