% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_info.R
\name{boundaries_info}
\alias{boundaries_info}
\title{Retrieve a list of ZCTAs with their spacial coordinates and a computed haversine distance in miles from the map center.}
\usage{
boundaries_info(path, map_center = NULL)
}
\value{
list of US ZCTAs as data.frame
}
\description{
Retrieve a list of ZCTAs with their spacial coordinates and a computed haversine distance in miles from the map center.
}
\examples{

}

