% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries_labels.R
\name{boundaries_labels}
\alias{boundaries_labels}
\title{boundaries_labels within 25miles radius of the center of the map. remove close zipcodes.}
\usage{
boundaries_labels(US_ZCTA)
}
\value{

}
\description{
boundaries_labels within 25miles radius of the center of the map. remove close zipcodes.
}
\examples{

}

