% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-helper.R
\name{convertImageListToDataFrame}
\alias{convertImageListToDataFrame}
\title{Reads image list from file and converts it into the needed data frame format}
\usage{
convertImageListToDataFrame(file)
}
\arguments{
\item{file}{path to json file}
}
\value{
data frame with relevant elements from json
}
\description{
Reads image list from file and converts it into the needed data frame format
}
