% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09-convertJsonToDataFrame.R
\name{convertJsonToDataFrame}
\alias{convertJsonToDataFrame}
\title{Reads json from file and converts it into the needed data frame format}
\usage{
convertJsonToDataFrame(file)
}
\arguments{
\item{file}{path to json file}
}
\value{
data frame with relevant elements from json
}
\description{
Reads json from file and converts it into the needed data frame format
}
