% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_map.R
\name{create_map}
\alias{create_map}
\title{Add together two numbers.}
\usage{
create_map(data, shape_zcta, shape_roads, zcta_names, map_center, dist_center,
  locations, legend_title, zoom)
}
\arguments{
\item{shape_roads}{road data}

\item{directory}{where do save the map png}

\item{mapname}{file name}

\item{legendtitle}{color bar legend title}

\item{metric}{heatmap metric}

\item{zipdata}{zip data}

\item{venue}{venue location}

\item{distvenue}{size of the map}

\item{zipnames}{zip codes}
}
\value{
a heat map ggplot.
}
\description{
Add together two numbers.
}

