% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-createQuestionnaireInputs.R
\name{createQuestionnaireInputs}
\alias{createQuestionnaireInputs}
\title{Create Inputs for Questionnaire}
\usage{
createQuestionnaireInputs(id, questions)
}
\arguments{
\item{id}{id of module}

\item{questions}{list of questions from json}
}
\description{
Create Inputs for Questionnaire
}
