% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-createVariableSelectionInputs.R
\name{createVariableSelectionInputs}
\alias{createVariableSelectionInputs}
\title{Create Inputs for Variable Selection}
\usage{
createVariableSelectionInputs(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Create Inputs for Variable Selection
}
