% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{CT}
\alias{CT}
\title{Create correspondence table object}
\usage{
CT(
  x,
  source = 1,
  target = 2,
  value = 3,
  source_name = "source",
  target_name = "target"
)
}
\arguments{
\item{x}{either in long format (data.table/frame with three cols: source, target, value), or in wide format (matrix) with rownames (source) and colnames (target)}

\item{source}{numeric: column of x where to find the source classification (only if x is data.table)}

\item{target}{numeric: column of x where to find the target classification (only if x is data.table)}

\item{value}{numeric: column of x where to find the value classification (only if x is data.table)}

\item{source_name}{character: name of source classification (only if x is matrix)}

\item{target_name}{character: name of target classification (only if x is matrix)}
}
\value{
an CT object (inherits from data.table and data.frame)
}
\description{
Create correspondence table object
}
