% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataExport.R
\name{dataExport}
\alias{dataExport}
\title{Server function of data export module}
\usage{
dataExport(input, output, session, df, filename = "data")
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{df}{data frame to be exported}

\item{filename}{name of file}
}
\description{
Server function of data export module
}
