% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataExport.R
\name{dataExportButton}
\alias{dataExportButton}
\title{UI function of data export button}
\usage{
dataExportButton(id, title = "Export Data")
}
\arguments{
\item{id}{id of module}

\item{title}{title of button}
}
\description{
UI function of data export button
}
