\name{datatests}
\alias{datatests}
\docType{data}
\title{
Data for Examples
}
\description{
This object contains data for examples.
}
\format{
A list with 18 variables:
  \describe{
    \item{mRNAProfile}{
      a dataframe representing mRNAs expression profile in breast cancer.It includes 10 rows standing for 10 genes and 112 columns standing for 112 samples.Its rownames are gene symbols.
    }\item{proteinProfile}{
      a dataframe representing protein expression profile in breast cancer.It includes 10 rows standing for 10 genes and 112 columns standing for 112 samples.Its rownames are gene symbols.
    }
  }
}
\details{
All expression data is about breast cancer generated from CPTAC project.
}
\keyword{datasets}
