% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-downloadSessionModule.R
\name{downloadSessionUI}
\alias{downloadSessionUI}
\title{Create Inputs for Variable Selection}
\usage{
downloadSessionUI(id)
}
\arguments{
\item{id}{id of module}
}
\description{
Create Inputs for Variable Selection
}
