% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ellie}
\alias{ellie}
\title{Elephant seal Argos satellite data (2 individuals)}
\format{
A tibble with 709 rows and 5 columns:
\describe{
\item{id}{individual identifier}
\item{date}{time of fix, POSIXct in UTC}
\item{lc}{Argos location class, from most to least accurate: 3, 2, 1, 0, A, B, Z}
\item{lon}{longitude in decimal degrees, WGS84}
\item{lat}{latitude in decimal degrees, WGS84}
}
}
\usage{
ellie
}
\description{
Example elephant seal Argos tracking data. Data were sourced from
the Integrated Marine Observing System (IMOS) - IMOS is supported by the
Australian Government through the National Collaborative Research Infrastructure
Strategy and the Super Science Initiative.
}
\keyword{data}
