% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-tableModule.R
\name{formatCaption}
\alias{formatCaption}
\title{Create a caption for a table from a title object}
\usage{
formatCaption(title_format)
}
\arguments{
\item{title_format}{named list with title text and format arguments}
}
\value{
(HTML) caption for a table
}
\description{
Create a caption for a table from a title object
}
