% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maybe.R
\name{from_just}
\alias{from_just}
\title{Unwraps a 'Just' value}
\usage{
from_just(x)
}
\arguments{
\item{x}{a \code{maybe} value (defined in the \strong{maybe} package)}
}
\value{
If \code{x} is a 'Just' \code{maybe} value, then the function 
  unwraps the value. Otherwise, an error is thrown.
}
\description{
Unwraps the content of a 'Just' \code{maybe} value.
}
\examples{
library(maybe)
from_just(just(9))
}
