% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gannets}
\alias{gannets}
\title{GPS tracks of immature northern gannets from the Bass Rock}
\format{
A tibble with 8158 rows and 5 columns:
\describe{
\item{id}{tag identifier, character}
\item{trip}{foraging trip identifier, character, unique across birds}
\item{datetime_utc}{time of fix, POSIXct in UTC}
\item{lon}{longitude in decimal degrees, WGS84}
\item{lat}{latitude in decimal degrees, WGS84}
}
}
\source{
Grecian et al. (2018) \doi{10.1098/rsif.2018.0084}
}
\usage{
gannets
}
\description{
GPS tracks of three immature northern gannets (\emph{Morus bassanus}) tagged at
the Bass Rock, Firth of Forth, in summer 2015. A subset of a larger
dataset, chosen so that the three birds between them span the North Sea
from Shetland to Dutch waters.
}
\keyword{data}
