% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapr-package.R
\docType{data}
\name{gbif_eg1}
\alias{gbif_eg1}
\title{Example dataset: output from call to \code{rgbif::occ_search()}}
\format{
A data frame with 50 rows and 101 variables
}
\description{
A dataset with 50 rows, and 101 columns, from the query:
\code{rgbif::occ_search(scientificName = "Puma concolor", limit = 100)}
}
\details{
See \code{inst/ignore/datasets.R} for the code to prepare this dataaset
}
\keyword{datasets}
