% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{is_coherent.CT}
\alias{is_coherent.CT}
\title{Title}
\usage{
\method{is_coherent}{CT}(x, tol = 1e-06, return_incoherent_sectors = FALSE)
}
\arguments{
\item{return_incoherent_sectors}{}
}
\value{

}
\description{
Title
}
