% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{iter_pairs}
\alias{iter_pairs}
\title{Iter Pairs
Iterate over subsequent pairs of elements}
\usage{
iter_pairs(vec)
}
\arguments{
\item{vec}{A vector to be iter-paired}
}
\value{
A list of 2-element lists
}
\description{
Iter Pairs
Iterate over subsequent pairs of elements
}
\details{
iter_pairs returns a list of adjacent pairs. Elements
from the original list paired like so: (1,2), (2,3), ..., (n-1, n)
}
