% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_us_zcta.R
\name{load_us_zcta}
\alias{load_us_zcta}
\title{Retrieve a list of ZCTAs with their spacial coordinates and a computed haversine distance in miles from the map center.}
\usage{
load_us_zcta(us_zcta_path, map_center = NULL)
}
\value{
list of US ZCTAs as data.frame
}
\description{
Retrieve a list of ZCTAs with their spacial coordinates and a computed haversine distance in miles from the map center.
}

