% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{make_CT_sankeystyle}
\alias{make_CT_sankeystyle}
\title{Transforms correspondence table into a format suitable for plotly sankey plot}
\usage{
make_CT_sankeystyle(x, source = 1, target = 2, value = 3)
}
\arguments{
\item{x}{a correspondence table in long format (data.table)}

\item{source}{location (numeric) or name (character) of column of source classification}

\item{target}{location (numeric) or name (character) of column of target classification}

\item{value}{location (numeric) or name (character) of value column}
}
\value{

}
\description{
Transforms correspondence table into a format suitable for plotly sankey plot
}
