% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_ggplot.R
\name{map_ggplot}
\alias{map_ggplot}
\title{ggplot2 visualization of species occurences}
\usage{
map_ggplot(df, zoom = 5, point_color = "#86161f")
}
\arguments{
\item{df}{Input \code{data.frame}}

\item{zoom}{zoom level for map. Adjust depending on how your data look.}

\item{point_color}{Default color of your points}
}
\description{
ggplot2 visualization of species occurences
}
\examples{
\dontrun{
ecoengine_data <- occ(query = 'Lynx rufus californicus', from = 'ecoengine', limit=100)
map_ggplot(ecoengine_data)
gbif_data <- occ(query = 'Accipiter striatus', from = 'gbif', limit=100)
map_ggplot(gbif_data)
bison_data <- occ(query = 'Accipiter striatus', from = 'bison', limit=100)
map_ggplot(bison_data)
}
}

