% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-mapPanelModule.R
\name{mapPanelServer}
\alias{mapPanelServer}
\title{Server function of mapPanel module}
\usage{
mapPanelServer(id)
}
\arguments{
\item{id}{module id}
}
\description{
Server function of mapPanel module
}
