% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-observeEnableActionButton.R
\name{observeEnableActionButton}
\alias{observeEnableActionButton}
\title{Enables or disables the action button depending on the availability of the input variables}
\usage{
observeEnableActionButton(input, image_list, questionnaire)
}
\arguments{
\item{input}{input object from server function}

\item{image_list}{reactive image list}

\item{questionnaire}{reactive questionnaire}
}
\description{
Enables or disables the action button depending on the availability of the input variables
}
