% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-observeShowAndHideInputs.R
\name{observeShowAndHideInputs}
\alias{observeShowAndHideInputs}
\title{Observer to show and hide inputs}
\usage{
observeShowAndHideInputs(
  input,
  output,
  session,
  uploadedZip,
  image_list,
  questionnaire,
  id
)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{uploadedZip}{reactive value of uploaded zip file}

\item{image_list}{reactive value of image list}

\item{questionnaire}{reactive value of questionnaire}

\item{id}{module id}
}
\description{
Observer to show and hide inputs
}
