% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-observeShowPlot.R
\name{observeShowPlot}
\alias{observeShowPlot}
\title{Observer to show plot when display button is clicked}
\usage{
observeShowPlot(
  input,
  output,
  session,
  image_list,
  questionnaire,
  title_format
)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{image_list}{reactive image list}

\item{questionnaire}{reactive questionnaire}

\item{title_format}{reactive list with title text and format arguments}
}
\description{
Observer to show plot when display button is clicked
}
