% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-observeUploadedZip.R
\name{observeUploadedZip}
\alias{observeUploadedZip}
\title{Observer for zip upload}
\usage{
observeUploadedZip(
  input,
  output,
  session,
  uploaded_zip,
  image_list,
  questionnaire,
  uploaded_inputs,
  upload_description,
  id
)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{uploaded_zip}{reactive value of uploaded zip file}

\item{image_list}{reactive value of image list}

\item{questionnaire}{reactive value of questionnaire}

\item{uploaded_inputs}{reactive value of uploaded inputs}

\item{upload_description}{reactive value of upload description}

\item{id}{module id}
}
\description{
Observer to show and hide inputs based on the zip file content. If a questionnaire is present,
the variable selection inputs are hidden and the questionnaire inputs are shown. If no
questionnaire is present, the variable selection inputs are shown and the questionnaire inputs
are hidden. The observer also extracts the questionnaire and image list from the zip file.
If existing, the observer extracts the notes and inputs from the zip file.
}
