% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/occdat_plot.R
\name{plot.occdat}
\alias{plot.occdat}
\title{Use base R plotting methods to visualize spocc data.}
\usage{
plot.occdat(x, ...)
}
\description{
Use base R plotting methods to visualize spocc data.
}

