% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plotModule.R
\name{plotServer}
\alias{plotServer}
\title{Server function of plot module}
\usage{
plotServer(id, path, file_type, title_format, variable = NULL, time = NULL)
}
\arguments{
\item{id}{id of the module}

\item{path}{path to the file}

\item{file_type}{file type of file}

\item{title_format}{reactive list with title text and format arguments}

\item{variable}{variable from user selection}

\item{time}{time from user selection}
}
\description{
Server function of plot module
}
