% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-observeShowPlot.R
\name{prepareImageListImage}
\alias{prepareImageListImage}
\title{Prepare infos for image from image list}
\usage{
prepareImageListImage(input, image_list)
}
\arguments{
\item{input}{input object from server function}

\item{image_list}{reactive image list}
}
\description{
Prepare infos for image from image list
}
