% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-observeShowPlot.R
\name{prepareQuestionnaireImage}
\alias{prepareQuestionnaireImage}
\title{Prepare infos for image from questionnaire}
\usage{
prepareQuestionnaireImage(input, questionnaire)
}
\arguments{
\item{input}{input object from server function}

\item{questionnaire}{reactive questionnaire}
}
\description{
Prepare infos for image from questionnaire
}
