% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{priority}
\alias{priority}
\title{Spec Priority}
\usage{
priority(x)
}
\arguments{
\item{x}{A spec object}
}
\description{
Spec Priority
}
\details{
Specifications must be processed in a certain order
otherwise transformed variables may be overwritten. The processing
of specs done in the following order:
\enumerate{
\item{Missing}
\item{Exceptions}
\item{Clamp}
\item{Range sorted by ascending lower boundary value}
}
}
