% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-radioButtonsModule.R
\name{radioButtonsUI}
\alias{radioButtonsUI}
\title{UI function of radioButtons}
\usage{
radioButtonsUI(id, choices, ...)
}
\arguments{
\item{id}{module id}

\item{choices}{choices for radioButtons}

\item{...}{further arguments passed to input}
}
\description{
UI function of radioButtons
}
