% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-selectizeInputModule.R
\name{selectizeInputUI}
\alias{selectizeInputUI}
\title{UI function of selectizeInput}
\usage{
selectizeInputUI(id, label, ...)
}
\arguments{
\item{id}{module id}

\item{label}{label}

\item{...}{further arguments passed to input}
}
\description{
UI function of selectizeInput
}
