% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_boundaries.R
\name{shape_boundaries}
\alias{shape_boundaries}
\title{shape_boundaries import and load a boundaries shapefile}
\usage{
shape_boundaries(shape_boundaries_dir, shape_boundaries_filename)
}
\description{
shape_boundaries import and load a boundaries shapefile
}
\examples{

}

