% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_roads.R
\name{shape_roads}
\alias{shape_roads}
\title{shape_roads import and load a Primary Roads shapefile}
\usage{
shape_roads(shape_roads_dir, shape_roads_filename)
}
\description{
shape_roads import and load a Primary Roads shapefile
}
\examples{

}

