% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-sliderInputModule.R
\name{sliderInputUI}
\alias{sliderInputUI}
\title{UI function of sliderInput}
\usage{
sliderInputUI(id, label, ...)
}
\arguments{
\item{id}{module id}

\item{label}{label}

\item{...}{further arguments passed to input}
}
\description{
UI function of sliderInput
}
