% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_clamp}
\alias{spec_clamp}
\title{Clamp Spec}
\usage{
spec_clamp(ll = -Inf, ul = Inf)
}
\arguments{
\item{ll}{lower limit}

\item{ul}{upper limit}
}
\description{
Clamp Spec
}
\details{
Unlike other specs, the clamp spec does not
have a relative ordering. It only limits the range of a
continuous variable to be within a certain upper and lower
limit.
}
