% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_exception}
\alias{spec_exception}
\title{Exception Spec}
\usage{
spec_exception(value, level)
}
\arguments{
\item{value}{single value to be constrained}

\item{level}{relative predicted ordering}
}
\description{
Exception Spec
}
