% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec.R
\name{spec_range}
\alias{spec_range}
\title{Range Spec}
\usage{
spec_range(rng = c(-Inf, Inf), linc = TRUE, rinc = FALSE, mono = 0,
  level)
}
\arguments{
\item{rng}{length-2 numeric vector specifying interval}

\item{linc}{boolean left boundary inclusive or exclusive}

\item{rinc}{boolean right boundary inclusive or exclusive}

\item{mono}{monotonicity constraint {-1, 0, 1}}

\item{level}{relative predicted ordering}
}
\description{
Range Spec
}
