% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spoccutils-package.R
\docType{package}
\name{spoccutils-package}
\alias{spoccutils}
\alias{spoccutils-package}
\title{Utility functions for use with the spocc package}
\description{
Utility functions for use with the spocc package
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

