% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-tableModule.R
\name{tableServer}
\alias{tableServer}
\title{Server function of table module}
\usage{
tableServer(id, df, title_format)
}
\arguments{
\item{id}{id of module}

\item{df}{data frame to be displayed in table}

\item{title_format}{reactive list with title text and format arguments}
}
\description{
Server function of table module
}
