% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-textFormatModule.R
\name{textFormatUI}
\alias{textFormatUI}
\title{UI function of textFormat module}
\usage{
textFormatUI(id, label, ...)
}
\arguments{
\item{id}{module id}

\item{label}{label}

\item{...}{further arguments passed to input}
}
\description{
UI function of textFormat module
}
