% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/us_zcta_names.R
\name{us_zcta_names}
\alias{us_zcta_names}
\title{us_zcta_names within 25miles radius of the center of the map. remove close zipcodes.}
\usage{
us_zcta_names(US_ZCTA)
}
\description{
us_zcta_names within 25miles radius of the center of the map. remove close zipcodes.
}

