% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visr-package.R
\docType{package}
\name{visr-package}
\alias{visr}
\alias{visr-package}
\title{Vizualize species occurrence data}
\description{
Vizualize species occurrence data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

