% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshaping.R
\name{CT_to_long}
\alias{CT_to_long}
\title{Title}
\usage{
CT_to_long(
  x,
  source_name = "source",
  target_name = "target",
  sort_by = "source"
)
}
\arguments{
\item{sort_by}{}
}
\value{

}
\description{
Title
}
