% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull.R
\name{hull}
\alias{hull}
\title{Add a convex hull to a map}
\usage{
hull(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{ignored}
}
\value{
Adds a convex hull to the plot. See \code{\link[grDevices]{chull}}
for info.
}
\description{
Add a convex hull to a map
}
\details{
Can be used with \code{\link{map_leaflet}}, \code{\link{map_plot}},
and \code{\link{map_ggplot}}. Other methods in this package may be supported
in the future.
}
\examples{
# map spocc output, here using a built in object
data(occdat_eg1)
map_plot(occdat_eg1, hull = TRUE)

# map rgbif output, here using a built in object
hull(map_ggplot(occdat_eg1))

}

