% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_ggmap.R
\name{map_ggmap}
\alias{map_ggmap}
\title{ggpmap visualization of species occurences}
\usage{
map_ggmap(x, zoom = 5, point_color = "#86161f", lon = "longitude",
  lat = "latitude", ...)
}
\arguments{
\item{x}{The data. An object of class \code{occdat}, \code{occdatind},
\code{gbif}, or \code{data.frame}. The package \pkg{spocc} needed for
the first two, and \pkg{rgbif} needed for the third. When \code{data.frame}
input, any number of columns allowed, but with at least the following:
name (the taxonomic name), latitude (in dec. deg.), longitude (in dec. deg.)}

\item{zoom}{zoom level for map. Adjust depending on how your data look.}

\item{point_color}{Default color of your points}

\item{lon, lat}{(character) Longitude and latitude variable names. Ignored
unless \code{data.frame} input to \code{x} parameter. We attempt to guess, but
if nothing close, we stop. Default: \code{longitude} and \code{latitude}}

\item{...}{Ignored}
}
\description{
ggpmap visualization of species occurences
}
\examples{
\dontrun{
## spocc
library("spocc")
ed <- occ(query = 'Lynx rufus californicus', from = 'ecoengine', limit=100)
map_ggmap(ed)
gd <- occ(query = 'Accipiter striatus', from = 'gbif', limit=100)
map_ggmap(gd)
bd <- occ(query = 'Accipiter striatus', from = 'bison', limit=100)
map_ggmap(bd)

## gbif
'xxxx'

## data.frame
df <- data.frame(name = c('Poa annua', 'Puma concolor', 'Foo bar'),
                 longitude = c(-120, -121, -121),
                 latitude = c(41, 42, 45), stringsAsFactors = FALSE)
map_ggpmap(df)
}
}

