% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_leaflet.R
\name{map_leaflet}
\alias{map_leaflet}
\title{Make an interactive map with Leaflet.js}
\usage{
map_leaflet(x, lon = "longitude", lat = "latitude", ...)
}
\arguments{
\item{x}{Input}

\item{lon}{Longitude variable name}

\item{lat}{Latitude variable name}

\item{...}{Further args to \code{\link{points}}}
}
\description{
Make an interactive map with Leaflet.js
}
\examples{
\dontrun{
## spocc
library("spocc")
(out <- occ(query='Accipiter striatus', from='gbif', limit=50, has_coords=TRUE))
### with class occdat
map_leaflet(out)
### with class occdatind
map_leaflet(out$gbif)

## rgbif
library("rgbif")
res <- occ_search(scientificName = "Puma concolor", limit = 100)
map_leaflet(res)

## data.frame
df <- data.frame(longitude = c(-120,-121), latitude = c(41, 42), stringsAsFactors = FALSE)
map_leaflet(df)
}
}

