% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Base R mapping}
\usage{
map_plot(x, ...)
}
\arguments{
\item{x}{Input, an object of class \code{occdat}}

\item{...}{Further args to \code{\link{points}}}
}
\description{
Base R mapping
}
\examples{
\dontrun{
## spocc
library("spocc")
(out <- occ(query='Accipiter striatus', from='gbif', limit=100, has_coords=TRUE))
map_plot(out)

## rgbif
library("rgbif")
res <- occ_search(scientificName = "Puma concolor", limit = 100)
map_plot(res)
}
}

