% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Use base R plotting method to visualize spocc data.}
\usage{
map_plot(x, ...)
}
\arguments{
\item{x}{Input, an object of class \code{occdat}}

\item{...}{Further args to \code{\link{points}}}
}
\description{
Use base R plotting method to visualize spocc data.
}

