% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_plot.R
\name{map_plot}
\alias{map_plot}
\title{Use base R plotting method to visualize spocc data.}
\usage{
map_plot(x, ...)
}
\arguments{
\item{x}{Input, an object of class \code{occdat}}

\item{...}{Further args to \code{\link{points}}}
}
\description{
Use base R plotting method to visualize spocc data.
}
\examples{
\dontrun{
library("spocc")
(out <- occ(query='Accipiter striatus', from='gbif', limit=100, has_coords=TRUE))
map_plot(out)
}
}

