% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapr-package.R
\docType{package}
\name{mapr-package}
\alias{mapr}
\alias{mapr-package}
\title{Vizualize species occurrence data}
\description{
Vizualize species occurrence data
}
\section{Many inputs}{

All functions take the following four kinds of inputs
\itemize{
 \item An object of class \code{occdat}, from the package \code{spocc}. An object of
 this class is composed of many objects of class \code{occdatind}
 \item An object of class \code{occdatind}, from the package \code{spocc}
 \item An object of class \code{gbif}, from the package \code{rgbif}
 \item An object of class \code{data.frame}. This data.frame can have any columns,
 but must include a column for taxonomic names (e.g., \code{name}), and for latitude
 and longitude (we guess your lat/long columns, starting with the default
 \code{latitude} and \code{longitude})
 \item An object of class \code{SpatialPoints}
 \item An object of class \code{SpatialPointsDatFrame}
}
}

\section{Mapping options}{

\itemize{
 \item \code{\link{map_plot}} - Base R plots
 \item \code{\link{map_ggplot}} - ggplot2 plots
 \item \code{\link{map_ggmap}} - ggplot2 plots with map layers
 \item \code{\link{map_leaflet}} - Leaflet.js interactive maps
 \item \code{\link{map_gist}} - Ineractive, shareable maps on GitHub Gists
}
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

