% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapr-package.R
\docType{package}
\name{mapr-package}
\alias{mapr}
\alias{mapr-package}
\title{Vizualize species occurrence data}
\description{
Vizualize species occurrence data
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}

