% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapr.R
\name{mapr}
\alias{mapr}
\title{mapr}
\usage{
mapr(dat, prj, buff)
}
\arguments{
\item{dat}{a data frame of observations (see details)}

\item{prj}{a PROJ.4 compatable projection for the region of interest *NOT*
WGS84. If dat inherits sf this will be scraped from the dataframe}

\item{buff}{a buffer to expand region of interest (specified in metres) DEFAULT
is 1e6 m (1000 km)}
}
\value{
\item{\code{world_shp}}{a projected shapefile for the region of
interest taken from Andy South's rworldmap package }
}
\description{
Generate a land shapefile for a region of interest specified by telemetry
data
}
\details{
Telemetry data is given either as an sf dataframe, or a tibble where each row
is an observed location and columns \describe{ \item{'id'}{individual animal
identifier,}\item{'date'}{observation time (POSIXct, GMT),} \item{'lon'}
{observed longitude,} \item{'lat'}{observed latitude,} \item{'...'}{other
columns will be ignored} }
}
\examples{
\dontrun{
require(tidyverse)
require(sf)
data(ellie)
prj <- '+proj=laea +lat_0=-60 +lon_0=70 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs'
world_shp <- mapr(ellie, prj, buff = 1e6)

ggplot() +
  geom_sf(aes(), data = world_shp) +
  geom_sf(aes(), data = st_as_sf(ellie, coords = c('lon', 'lat')) \%>\%
      st_set_crs('+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'))
}
}
