% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meshr.R
\name{meshr}
\alias{meshr}
\title{meshr}
\usage{
meshr(dat, prj, buff, keep)
}
\arguments{
\item{dat}{a data frame of observations (see details)}

\item{prj}{a PROJ.4 compatable projection for the region of interest *NOT*
WGS84}

\item{buff}{a buffer to expand region of interest specified in metres}

\item{keep}{the proportion of points to be retained - passed to rmapshaper::ms_simplify}
}
\value{
a list containing an inla mesh boundary for the region of interest
}
\description{
Generate an inla mesh nested boundary for a region of interest
specified by telemetry data
}
\details{
The telemetry data is given as a dataframe where each row is an observed
location and columns \describe{ \item{'id'}{individual animal identifier,}
\item{'date'}{observation time (POSIXct, GMT),} \item{'lon'}{observed
longitude,} \item{'lat'}{observed latitude,} \item{'...'}{other columns
will be ignored} }
}
\examples{
\dontrun{

require(tidyverse)
require(sf)
require(INLA)
require(inlabru)

data(ellie)
prj <- '+proj=laea +lat_0=-60 +lon_0=70 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs'

b <- meshr(ellie, prj, buff = 5e5, keep = 0.02)

mesh = inla.mesh.2d(boundary = b, max.edge = c(250000, 1000000), cutoff = 25000, max.n = 1000)

ggplot() +
  geom_sf(aes(), data = mapr(ellie, prj, buff = 1e6)) +
  inlabru::gg(mesh) +
  geom_sf(aes(), data = st_as_sf(ellie, coords = c('lon', 'lat')) \%>\%
    st_set_crs('+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'))
}
}
