% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-tableModule.R
\name{observeShowTable}
\alias{observeShowTable}
\title{Observer to show table when display button is clicked}
\usage{
observeShowTable(input, output, session, image_list, table_data)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{image_list}{reactive image list}

\item{table_data}{reactive table data}
}
\description{
Observer to show table when display button is clicked
}
