% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_to_df.R
\name{shapefile_to_df}
\alias{shapefile_to_df}
\title{shapefile_to_df import and load a boundaries shapefile}
\usage{
shapefile_to_df(shape_boundaries_dir, shape_boundaries_filename,
  boundary_identifier = NULL, show_data = FALSE)
}
\arguments{
\item{shape_boundaries_dir}{Unique boundary identifier. US boundary_identifier = "ZCTA5CE10". Canada boundary_identifier = "CFSAUID"}

\item{shape_boundaries_filename}{Unique boundary identifier. US boundary_identifier = "ZCTA5CE10". Canada boundary_identifier = "CFSAUID"}

\item{boundary_identifier}{Unique boundary identifier. US boundary_identifier = "ZCTA5CE10". Canada boundary_identifier = "CFSAUID"}
}
\value{
a spacial object as a data.frame
}
\description{
shapefile_to_df import and load a boundaries shapefile
}
\examples{

}

