% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-tableModule.R
\name{tableServer}
\alias{tableServer}
\title{Server function of table module}
\usage{
tableServer(id, df)
}
\arguments{
\item{id}{id of module}

\item{df}{data frame to be displayed in table}
}
\description{
Server function of table module
}
