% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umapR.R
\name{umapR}
\alias{umapR}
\title{R6 class representing an unordered map.}
\description{
A map is given by keys and values. The keys must be 
  some strings, the values can be any R objects.
}
\examples{

## ------------------------------------------------
## Method `umapR$new`
## ------------------------------------------------

umapR$new(
  keys = c("a", "b"), 
  values = list(c(1, 2), c(3, 4, 5))
)
# examples with duplicated keys:
umapR$new(
  keys = c("a", "a", "b"), 
  values = list(c(1, 2), c(3, 4), c(5, 6))
)
umapR$new(
  keys = c("a", "a", "b"), 
  values = list(c(1, 2), c(3, 4), c(5, 6)),
  duplicated = "join"
)
umapR$new(
  keys = c("a", "a", "b"), 
  values = list(c(1, 2), c(3, 4), c(5, 6)),
  duplicated = "separate"
)

## ------------------------------------------------
## Method `umapR$size`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$size()

## ------------------------------------------------
## Method `umapR$keys`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$keys()

## ------------------------------------------------
## Method `umapR$values`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$values()

## ------------------------------------------------
## Method `umapR$items`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$items()

## ------------------------------------------------
## Method `umapR$toList`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$toList()

## ------------------------------------------------
## Method `umapR$at`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$at("b")
map$at("x", stop_if_not_found = FALSE)

## ------------------------------------------------
## Method `umapR$extract`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b", "c"), 
  values = list(c(1, 2), c(3, 4, 5), c(6, 7))
)
map$extract(c("a", "c"))

## ------------------------------------------------
## Method `umapR$submap`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b", "c"), 
  values = list(c(1, 2), c(3, 4, 5), c(6, 7))
)
map$submap(c("a", "c"))

## ------------------------------------------------
## Method `umapR$has_key`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$has_key("b")
map$has_key("x")

## ------------------------------------------------
## Method `umapR$insert`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$insert("c", c(6, 7))
map
map$insert("a", c(8, 9))
map
map$insert("a", c(8, 9), replace = TRUE)
map

## ------------------------------------------------
## Method `umapR$erase`
## ------------------------------------------------

map <- umapR$new(
  keys = c("a", "b", "c"), 
  values = list(c(1, 2), c(3, 4, 5), c(6, 7))
)
map$erase("a")
map
map$erase(c("b", "c"))
map

## ------------------------------------------------
## Method `umapR$merge`
## ------------------------------------------------

map1 <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map2 <- umapR$new(
  keys = c("c", "d"), values = list(c(9, 8), c(7, 6))
)
map1$merge(map2)
map1

# `duplicated` example ####
map1 <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map1_copy1 <- map1$copy()
map1_copy2 <- map1$copy()
map1_copy3 <- map1$copy()
map2 <- umapR$new(
  keys = c("a", "d"), values = list(c(9, 8), c(7, 6))
)
map1_copy1$merge(map2)
map1_copy1

map1_copy2$merge(map2, duplicated = "join")
map1_copy2

map1_copy3$merge(map2, duplicated = "separate")
map1_copy3

## ------------------------------------------------
## Method `umapR$copy`
## ------------------------------------------------

map <- umapR$new(c("a", "b"), list(c(1,2), c(FALSE, TRUE)))
true_copy <- map$copy()
true_copy$erase("a")
map
naive_copy <- map
naive_copy$erase("a")
map
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{umapR$new()}}
\item \href{#method-print}{\code{umapR$print()}}
\item \href{#method-size}{\code{umapR$size()}}
\item \href{#method-keys}{\code{umapR$keys()}}
\item \href{#method-values}{\code{umapR$values()}}
\item \href{#method-items}{\code{umapR$items()}}
\item \href{#method-toList}{\code{umapR$toList()}}
\item \href{#method-at}{\code{umapR$at()}}
\item \href{#method-extract}{\code{umapR$extract()}}
\item \href{#method-submap}{\code{umapR$submap()}}
\item \href{#method-has_key}{\code{umapR$has_key()}}
\item \href{#method-insert}{\code{umapR$insert()}}
\item \href{#method-erase}{\code{umapR$erase()}}
\item \href{#method-merge}{\code{umapR$merge()}}
\item \href{#method-copy}{\code{umapR$copy()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Creates a new \code{umapR} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$new(keys, values, duplicated = "drop", checks = TRUE, ptr = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{the keys, a character vector without \code{NA} value}

\item{\code{values}}{the values, a list of R objects; \code{keys} and 
\code{values} must have the same length}

\item{\code{duplicated}}{the action to perform for duplicated keys, one of 
\code{"drop"}, \code{"join"}, or \code{"separate"}}

\item{\code{checks}}{Boolean, whether to check \code{keys} and \code{values}}

\item{\code{ptr}}{an external pointer; this is for internal use only}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{umapR} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{umapR$new(
  keys = c("a", "b"), 
  values = list(c(1, 2), c(3, 4, 5))
)
# examples with duplicated keys:
umapR$new(
  keys = c("a", "a", "b"), 
  values = list(c(1, 2), c(3, 4), c(5, 6))
)
umapR$new(
  keys = c("a", "a", "b"), 
  values = list(c(1, 2), c(3, 4), c(5, 6)),
  duplicated = "join"
)
umapR$new(
  keys = c("a", "a", "b"), 
  values = list(c(1, 2), c(3, 4), c(5, 6)),
  duplicated = "separate"
)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Show instance of a \code{umapR} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{ignored}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-size"></a>}}
\if{latex}{\out{\hypertarget{method-size}{}}}
\subsection{Method \code{size()}}{
Size of the map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer, the number of entries.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$size()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-keys"></a>}}
\if{latex}{\out{\hypertarget{method-keys}{}}}
\subsection{Method \code{keys()}}{
Get all keys.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The keys, a character vector.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$keys()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-values"></a>}}
\if{latex}{\out{\hypertarget{method-values}{}}}
\subsection{Method \code{values()}}{
Get all values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$values()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The values, a list of R objects.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$values()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-items"></a>}}
\if{latex}{\out{\hypertarget{method-items}{}}}
\subsection{Method \code{items()}}{
Get all entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$items()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The entries in a dataframe.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$items()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toList"></a>}}
\if{latex}{\out{\hypertarget{method-toList}{}}}
\subsection{Method \code{toList()}}{
Converts the map to a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$toList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$toList()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-at"></a>}}
\if{latex}{\out{\hypertarget{method-at}{}}}
\subsection{Method \code{at()}}{
Returns the value corresponding to the given key.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$at(key, stop_if_not_found = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a key (string)}

\item{\code{stop_if_not_found}}{a Boolean value, whether to stop if the key 
is not found, or to return \code{NaN}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The value corresponding to the key, an R object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$at("b")
map$at("x", stop_if_not_found = FALSE)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-extract"></a>}}
\if{latex}{\out{\hypertarget{method-extract}{}}}
\subsection{Method \code{extract()}}{
Extract submap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$extract(keys, inplace = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{some keys, a character vector; those which do not belong to 
the keys of the reference map will be ignored}

\item{\code{inplace}}{Boolean, whether to update the reference map or 
to return a new map}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{umapR} object if \code{inplace=FALSE}, 
  nothing otherwise.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b", "c"), 
  values = list(c(1, 2), c(3, 4, 5), c(6, 7))
)
map$extract(c("a", "c"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submap"></a>}}
\if{latex}{\out{\hypertarget{method-submap}{}}}
\subsection{Method \code{submap()}}{
Extract submap.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$submap(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{some keys, a character vector; those which do not belong to 
the keys of the reference map will be ignored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{umapR} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b", "c"), 
  values = list(c(1, 2), c(3, 4, 5), c(6, 7))
)
map$submap(c("a", "c"))
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-has_key"></a>}}
\if{latex}{\out{\hypertarget{method-has_key}{}}}
\subsection{Method \code{has_key()}}{
Checks whether a key exists in a map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$has_key(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a key (string)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Boolean value.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$has_key("b")
map$has_key("x")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-insert"></a>}}
\if{latex}{\out{\hypertarget{method-insert}{}}}
\subsection{Method \code{insert()}}{
Insert a new entry.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$insert(key, value, replace = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{a key (string)}

\item{\code{value}}{a value (R object)}

\item{\code{replace}}{Boolean, whether to replace the value if the key is 
already present}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, this updates the map.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map$insert("c", c(6, 7))
map
map$insert("a", c(8, 9))
map
map$insert("a", c(8, 9), replace = TRUE)
map
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-erase"></a>}}
\if{latex}{\out{\hypertarget{method-erase}{}}}
\subsection{Method \code{erase()}}{
Erase some entries.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$erase(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{some keys, a character vector}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, this updates the map.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(
  keys = c("a", "b", "c"), 
  values = list(c(1, 2), c(3, 4, 5), c(6, 7))
)
map$erase("a")
map
map$erase(c("b", "c"))
map
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-merge"></a>}}
\if{latex}{\out{\hypertarget{method-merge}{}}}
\subsection{Method \code{merge()}}{
Merge with another map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$merge(map, duplicated = "drop")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{map}}{a \code{umapR} object}

\item{\code{duplicated}}{the action to perform if the reference map 
and \code{map} have some identical keys, one of 
\code{"drop"}, \code{"join"}, or \code{"separate"}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, this updates the reference map.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map1 <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map2 <- umapR$new(
  keys = c("c", "d"), values = list(c(9, 8), c(7, 6))
)
map1$merge(map2)
map1

# `duplicated` example ####
map1 <- umapR$new(
  keys = c("a", "b"), values = list(c(1, 2), c(3, 4, 5))
)
map1_copy1 <- map1$copy()
map1_copy2 <- map1$copy()
map1_copy3 <- map1$copy()
map2 <- umapR$new(
  keys = c("a", "d"), values = list(c(9, 8), c(7, 6))
)
map1_copy1$merge(map2)
map1_copy1

map1_copy2$merge(map2, duplicated = "join")
map1_copy2

map1_copy3$merge(map2, duplicated = "separate")
map1_copy3
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-copy"></a>}}
\if{latex}{\out{\hypertarget{method-copy}{}}}
\subsection{Method \code{copy()}}{
Copy the reference map.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{umapR$copy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A copy of the reference map.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{map <- umapR$new(c("a", "b"), list(c(1,2), c(FALSE, TRUE)))
true_copy <- map$copy()
true_copy$erase("a")
map
naive_copy <- map
naive_copy$erase("a")
map
}
\if{html}{\out{</div>}}

}

}
}
